feat: split wws code in multiple crates
