+++ .github/workflows/artifacts.yml
@@ -11,7 +11,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Test
-      run: cargo test
+      run: cargo test --all
   build:
     strategy:
       matrix:
