+++ Cargo.toml
@@ -1,40 +1,42 @@
 [package]
 name = "wasm-workers-server"
 description = "A WebAssembly framework to develop and run serverless applications anywhere"
-version = "1.0.1"
-edition = "2021"
-license = "Apache-2.0"
 homepage = "https://workers.wasmlabs.dev/"
 documentation = "https://workers.wasmlabs.dev/docs/intro"
-repository = "https://github.com/vmware-labs/wasm-workers-server/"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+# Common package definition for all crates / kits
+[workspace.package]
+version = "1.0.1"
+edition = "2021"
+authors = [ "Wasm Labs <https://wasmlabs.dev>" ]
+license = "Apache-2.0"
+repository = "https://github.com/vmware-labs/wasm-workers-server/"
 
 [[bin]]
 name = "wws"
 path = "src/main.rs"
 
+# Main binary (wws CLI dependencies)
 [dependencies]
-wasmtime = "6.0.1"
-wasmtime-wasi = "6.0.1"
-anyhow = "1.0.66"
-blake3 = "1.3.3"
-wasi-common = "6.0.1"
-actix-web = "4"
+anyhow = { workspace = true }
+actix-web = { workspace = true }
 actix-files = "0.6.2"
-lazy_static = "1.4.0"
+lazy_static = { workspace = true }
 env_logger = "0.9.0"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.85"
-wax = "0.5.0"
-toml = "0.7.0"
-base64 = "0.21.0"
 clap = { version = "4.0.10", features = ["derive"] }
-regex = "1"
-url = "2.3.1"
-reqwest = { version = "0.11" }
-sha256 = "1.1.1"
 prettytable-rs = "0.10.0"
+wws-config = { workspace = true }
+wws-data-kv = { workspace = true }
+wws-router = { workspace = true }
+wws-store = { workspace = true }
+wws-runtimes = { workspace = true }
+wws-runtimes-manager = { workspace = true }
+wws-worker = { workspace = true }
 
 [target.x86_64-unknown-linux-musl.dependencies]
 openssl = { version = "=0.10.45", features = ["vendored"] }
@@ -44,6 +46,13 @@ openssl = { version = "=0.10.45", features = ["vendored"] }
 
 [workspace]
 members = [
+  "crates/config",
+  "crates/data-kv",
+  "crates/router",
+  "crates/runtimes",
+  "crates/runtimes-manager",
+  "crates/store",
+  "crates/worker",
   "kits/rust",
   "kits/rust/worker",
   "kits/javascript"
@@ -55,3 +64,21 @@ exclude = [
   "examples/rust-kv",
   "examples/rust-params"
 ]
+
+[workspace.dependencies]
+actix-web = "4"
+anyhow = "1.0.66"
+lazy_static = "1.4.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.85"
+toml = "0.7.0"
+wws-config = { path = "./crates/config" }
+wws-runtimes = { path = "./crates/runtimes" }
+wws-runtimes-manager = { path = "./crates/runtimes-manager" }
+wws-data-kv = { path = "./crates/data-kv" }
+wws-router = { path = "./crates/router" }
+wws-store = { path = "./crates/store" }
+wws-worker = { path = "./crates/worker" }
+wasmtime = "6.0.1"
+wasmtime-wasi = "6.0.1"
+wasi-common = "6.0.1"
