+++ crates/config/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "wws-config"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+toml = { workspace = true }
+wws-runtimes-manager = { workspace = true }
