+++ crates/data-kv/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "wws-data-kv"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
