+++ crates/router/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "wws-router"
+version = { workspace = true }
+edition = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+
+[lib]
+doctest = false
+
+[dependencies]
+lazy_static = { workspace = true }
+wws-config = { workspace = true }
+wws-store = { workspace = true }
+wws-runtimes-manager = { workspace = true }
+wws-worker = { workspace = true }
+regex = "1"
+wax = "0.5.0"
