+++ crates/router/src/route.rs
@@ -1,7 +1,6 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{config::Config as ProjectConfig, workers::config::Config, workers::Worker};
 use lazy_static::lazy_static;
 use regex::Regex;
 use std::{
@@ -9,6 +8,8 @@ use std::{
     fs,
     path::{Component, Path, PathBuf},
 };
+use wws_config::Config as ProjectConfig;
+use wws_worker::{config::Config, Worker};
 
 lazy_static! {
     static ref PARAMETER_REGEX: Regex = Regex::new(r"\[\w+\]").unwrap();
