+++ crates/runtimes-manager/src/fetch.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::runtimes::metadata::Checksum;
+use crate::metadata::Checksum;
 use anyhow::Result;
 use reqwest::header::USER_AGENT;
