+++ crates/runtimes/src/modules/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2022 VMware, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod external;
+pub(crate) mod javascript;
+pub(crate) mod native;
