+++ crates/worker/src/io.rs
