--- src/data/README.md
@@ -1,7 +1,0 @@
-# Data
-
-Data are a way to inject and read external information in the wasm modules. These data services reads the data from a given location and injects the data to the worker. Others allow workers to write data.
-
-## List
-
-* KV: provides a generic in-memory Key-Value store.
