--- src/router/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2022 VMware, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-mod files;
-mod route;
-mod routes;
-
-pub(crate) use routes::Routes;
