--- src/runtimes/mod.rs
@@ -1,7 +1,0 @@
-// Copyright 2022 VMware, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod manager;
-pub mod metadata;
-mod modules;
-pub mod runtime;
