--- src/runtimes/modules/mod.rs
@@ -1,6 +1,0 @@
-// Copyright 2022 VMware, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub(super) mod external;
-pub(super) mod javascript;
-pub(super) mod native;
