--- src/workers/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2022 VMware, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod config;
-pub mod wasm_io;
-mod worker;
-
-pub(crate) use worker::Worker;
